feat: change metadata to nonfile, make work with single input, and write to given output path.
